requirements.txt
fastapi
sqlalchemy
asyncpg
pydantic
pydantic-settings
uvicorn
alembic
requests
python-dotenv